+++ SCOTUS/byspkr/align_spkrdvec.sh
