+++ SCOTUS/byspkr/process_spkrdvec.sh
