+++ SCOTUS/dvec_bycase.py
@@ -27,106 +27,12 @@ import torch
 import json
 import csv
 import sys
-
-#assumes SVE is located in home dir
-sys.path.append("~/SpeakerVerificationEmbedding/src")
 sys.path.append("./SpeakerVerificationEmbedding/src")
-sys.path.append("./src")
 from hparam import hparam_SCOTUS as hp
 from speech_embedder_net import SpeechEmbedder
 from VAD_segments import VAD_chunk
-    
-
-def concat_segs(times, segs):
-    #Concatenate continuous voiced segments
-    concat_seg = []
-    seg_concat = segs[0]
-    hold_times = []
-    t0 = 0
-    for i in range(0, len(times)-1):
-        if times[i][1] == times[i+1][0]:
-            seg_concat = np.concatenate((seg_concat, segs[i+1]))
-        else:
-            hold_times.append((t0, times[i][1]))
-            t0 = times[i+1][0]
-            concat_seg.append(seg_concat)
-            seg_concat = segs[i+1]
-    else:
-        concat_seg.append(seg_concat)
-        hold_times.append((t0, times[-1][1]))
-    return concat_seg, hold_times
-
-def get_STFTs(segs,htemp):
-    #Get 240ms STFT windows with 50% overlap
-    sr = hp.data.sr
-    STFT_frames = []
-    STFT_labels = []
-    idx = 0
-    for seg in segs:
-        S = librosa.core.stft(y=seg, n_fft=hp.data.nfft,
-                              win_length=int(hp.data.window * sr), hop_length=int(hp.data.hop * sr))
-        S = np.abs(S)**2
-        mel_basis = librosa.filters.mel(sr, n_fft=hp.data.nfft, n_mels=hp.data.nmels)
-        S = np.log10(np.dot(mel_basis, S) + 1e-6)           # log mel spectrogram of utterances
-        for j in range(0, S.shape[1], int(.12/hp.data.hop)):
-            if j + 24 < S.shape[1]:
-                STFT_frames.append(S[:,j:j+24])
-                STFT_labels.append(htemp[idx])
-            else:
-                break
-        idx+=1
-    return STFT_frames, STFT_labels
-    
-def align_embeddings(embeddings, labs):
-    partitions = []
-    start = 0
-    end = 0
-    j = 1
-    for i, embedding in enumerate(embeddings):
-        if (i*.12)+.24 < j*.401:
-            end = end + 1
-        else:
-            partitions.append((start,end))
-            start = end
-            end = end + 1
-            j += 1
-    else:
-        partitions.append((start,end))
-    avg_embeddings = np.zeros((len(partitions),256))
-    emb_labels = []
-    for i, partition in enumerate(partitions):
-        emb_lab = labs[partition[0]:partition[1]]
-        if  len(set(emb_lab))>1:
-          continue
-        else:
-          avg_embeddings[i] = np.average(embeddings[partition[0]:partition[1]],axis=0) 
-          emb_labels.append(emb_lab)        
-    return avg_embeddings[0:len(emb_labels)], emb_labels
-
-def align_times(casetimelist, hold_times, spkr_dict):
-  htemp = []
-  _, endtime, endspkr = casetimelist[-1]
-  for h in hold_times:
-    append = False
-    for c in casetimelist:
-      if h[1]<c[1]:
-        if h[1]>=c[0]:
-          spkr_name = c[2]
-          htemp.append(spkr_dict[spkr_name])
-          append = True
-        else:
-          continue
-      else:
-        continue
-    if not append and h[1]!=hold_times[-1][1]:
-      print('value not appended in loop')
-      print(h)
-  htemp.append(spkr_dict[endspkr])
-  return htemp
-  
+from utils import concat_segs, get_STFTs, align_embeddings, align_times
  
- 
-    
 #-----------------------------
     
 #initialize SpeechEmbedder
@@ -159,7 +65,7 @@ for i, path in enumerate(case_path):
           spkr_dict[spkr] = cnt
           cnt+=1
           if cnt>=20:
-            print('ERROR: NEED MORE JUDGE ROOM')
+            print("ERROR: NEED MORE JUDGE ROOM")
       else:
         if spkr not in spkr_dict:
           spkr_dict[spkr] = label
@@ -167,13 +73,17 @@ for i, path in enumerate(case_path):
       filetimelist.append((float(t0),float(t1),spkr))
     casetimedict[file[:-4]] = filetimelist
 
-
+#save spkr label dictionary
+#PATH MANUAL (will adapt in full run)
+with open('/scratch/jt2565/sco50/info/spkrs.json', 'w') as outfile:  
+    json.dump(spkr_dict, outfile) 
 
 fold = hp.data.save_path
 cut_div = 4
 train_sequences = []
 train_cluster_ids = []
-print('starting generation')
+print("starting generation")
+
 for i, path in enumerate(case_path):
   
   file = path.split('/')[-1]
@@ -185,9 +95,9 @@ for i, path in enumerate(case_path):
     times, segs = VAD_chunk(2, path)
     concat_seg, ht = concat_segs(times, segs)
     htemp = align_times(casetimedict[file[:-4]], ht, spkr_dict)
-    if len(ht)!=len(htemp):
-        print('bad time diarization', len(ht), len(htemp))
-        continue
+    if hp.data.verbose:
+      print(len(ht))
+      print(len(htemp))
     STFT_frames, STFT_labels = get_STFTs(concat_seg, htemp)
     STFT_frames = np.stack(STFT_frames, axis=2)
     STFT_frames = torch.tensor(np.transpose(STFT_frames, axes=(2,1,0)))
@@ -200,13 +110,16 @@ for i, path in enumerate(case_path):
         STFT_samp = STFT_frames[t0:t0+cut, :, :]
       else:
         STFT_samp = STFT_frames[t0:, :, :]
+      if verbose:
+        print(STFT_samp.size())
       #process slice
-      STFT_samp = STFT_samp.to(hp.device)
+      STFT_samp = STFT_samp.to(device)
       emb = embedder_net(STFT_samp)
       temp_emb.append(emb.detach().cpu().numpy())
       t0+=cut
 
     embeddings = np.concatenate(temp_emb, axis=0)
+    print(embeddings.shape, len(STFT_labels))
     aligned_embeddings, aligned_labels = align_embeddings(embeddings, STFT_labels)
     train_sequences.append(aligned_embeddings)
     train_cluster_ids.append(aligned_labels)
@@ -214,8 +127,10 @@ for i, path in enumerate(case_path):
     np.save(fold+file[:-4]+'_id', aligned_labels)
     print('values appended')
 
-np.save('/scratch/jt2565/train_seq', train_sequences)
-np.save('/scratch/jt2565/train_clus', train_cluster_ids)
+full_set = False
+if full_set:
+    np.save('/scratch/jt2565/train_seq', train_sequences)
+    np.save('/scratch/jt2565/train_clus', train_cluster_ids)
