+++ src/importcheck.py
@@ -26,7 +26,7 @@ from hparam import hparam_SCOTUS as hp
 from VAD_segments import VAD_chunk
 from utils import get_centroids, get_cossim, calc_loss
 from speech_embedder_net import SpeechEmbedder, GE2ELoss, get_centroids, get_cossim
-from utils import concat_segs, get_STFTs, align_embeddings
+from utils import concat_segs, get_STFTs, align_embeddings, align_times
 print("imports complete")
 
 print("initialzie speech embedder")
